# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farah API Documentation",
        "description": "توثيق واجهات برمجة التطبيقات (APIs) لمنصة فرح لتطبيق Flutter",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://farahbackend-production.up.railway.app",
            "description": "سيرفر الإنتاج (Railway Production)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "السيرفر المحلي (Local Server)"
        }
    ],
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "جلب قائمة الخدمات المفضلة للمستخدم",
                "description": "يسترجع قائمة بجميع الخدمات التي قام المستخدم المسجل بإضافتها إلى المفضلة.",
                "operationId": "bc433e786ab5a0bec81c53c578e0b6aa",
                "responses": {
                    "200": {
                        "description": "تم جلب المفضلة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "service_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "صالة أورنا"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1400
                                                    },
                                                    "currency": {
                                                        "type": "string",
                                                        "example": "ILS"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "services/hall_orna.jpg",
                                                        "nullable": true
                                                    },
                                                    "rating_avg": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 4.8
                                                    },
                                                    "reviews_count": {
                                                        "type": "integer",
                                                        "example": 19
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "غزة - الرمال - شمال مطعم التايلندي"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "غزة"
                                                    },
                                                    "is_favorited": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "صالات"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "provider": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "صالة أورنا"
                                                            },
                                                            "avatar": {
                                                                "type": "string",
                                                                "example": "providers/orna.jpg",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح (يرجى تسجيل الدخول)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites/toggle/{service_id}": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "إضافة أو إزالة خدمة من المفضلة (Toggle)",
                "description": "يقوم بالتبديل بين إضافة الخدمة إلى المفضلة أو حذفها في حال كانت موجودة مسبقاً.",
                "operationId": "ec6c68a63bebfe2743dd6acf0bc0b770",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "معرّف الخدمة المطلوب إضافتها أو إزالتها",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم تحديث حالة المفضلة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "is_favorited": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تمت إضافة الخدمة إلى المفضلة بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "الخدمة غير موجودة"
                    },
                    "401": {
                        "description": "غير مصرح (يرجى تسجيل الدخول)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "أدخل التوكن بهذا الشكل: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}